llm-sql-copilot/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0